--- src/components/utils.js
@@ -1,44 +1,0 @@
-function closeByEscape(e) {
-    if (e.key === "Escape") {
-        const openedPopup = document.querySelector(".popup_opened");
-        closePopup(openedPopup);
-    }
-}
-
-/* Open popup */
-export function openPopup(popup) {
-    popup.classList.add('popup_opened');
-    document.addEventListener("keydown", closeByEscape);
-}
-
-/* Close popup */
-export function closePopup(popup) {
-    popup.classList.remove('popup_opened');
-    document.removeEventListener("keydown", closeByEscape); 
-}
-
-function renderLoading(isLoading, button, buttonText = 'Сохранить', loadingText = 'Сохранение...') {
-    if (isLoading) {
-        button.textContent = loadingText;
-    } else {
-        button.textContent = buttonText;
-    }
-}
-
-export function handleSubmit(request, evt, loadingText = "Сохранение...") {
-    evt.preventDefault();
-
-    const submitButton = evt.submitter;
-    const initialText = submitButton.textContent;
-    renderLoading(true, submitButton, initialText, loadingText);
-    request()
-        .then(() => {
-            evt.target.reset();
-        })
-        .catch((err) => {
-            console.error(`Ошибка: ${err}`);
-        })
-        .finally(() => {
-            renderLoading(false, submitButton, initialText);
-        });
-}
